refactor: enums 패키지명 constants로 수정
