+++ README.md
@@ -14,7 +14,7 @@
   - camp.nextstep.edu.missionutils.Console의 readLine()을 사용
 
 ### 기본 구분자로 숫자 추출 및 계산
-- [x] 쉼표(,) 또는 콜론(:)을 구분자로 가지는 문자열을 입력받아 숫자를 추출한다. 
+- [x] 쉼표(,) 또는 콜론(:)을 구분자로 가지는 문자열을 입력받아 숫자를 추출한다.
 - [x] 추출된 숫자들을 모두 더한 결과를 반환한다.
 
 ### 커스텀 구분자 추가 지정
@@ -26,6 +26,7 @@
 - [x] 입력 문자열에 음수나 0이 포함된 경우 "음수나 0은 입력할 수 없습니다."라는 메시지를 출력한다.
 - [x] 입력 문자열에 숫자가 아닌 값이 포함된 경우 "유효하지 않은 숫자 형식입니다."라는 메시지를 출력한다.
 - [x] 빈 문자열이나 null 입력 시 "입력값이 비어있습니다."라는 메시지를 출력한다.
+- [x] 구분자가 연속으로 입력되었을 경우 "구분자는 연속으로 입력할 수 없습니다."라는 메시지를 출력한다.
 - [x] 커스텀 구분자에 숫자가 들어갈 경우 "커스텀 구분자에 숫자는 입력할 수 없습니다."라는 메시지를 출력한다.
 - [x] 커스텀 구분자에 여러가지 구분자가 들어갈 경우 "커스텀 구분자는 한가지만 지정할 수 있습니다."라는 메시지를 출력한다.
 
@@ -53,6 +54,6 @@
 - view
   - InputView
   - OutputView
-- enums
+- constants
   - ExceptionMessage
   - Message
