+++ src/main/java/calculator/constants/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package calculator.enums;
+package calculator.constants;
 
 public class ExceptionMessage {
 	public static final String INPUT_EMPTY = "입력값이 비어있습니다.";
