+++ src/main/java/calculator/constants/Message.java
@@ -1,4 +1,4 @@
-package calculator.enums;
+package calculator.constants;
 
 public class Message {
 	public static final String START_MESSAGE = "계산기가 시작되었습니다.\n사용자는 덧셈할 문자열을 입력해 주세요.";
